refactor: apply TypeScript on TokenInput
