+++ docs/index.html
@@ -7,7 +7,7 @@
 <meta name="description" content="">
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet" />
-<script defer src="bundle.js?7c5dcf415b8be22748c4"></script></head>
+<script defer src="bundle.js?cbae513f706c612e2290"></script></head>
 <body>
 <div id="container"></div>
 <script src="https://polyfill.io/v3/polyfill.min.js"></script>
