+++ lib/index.js
@@ -1,3 +1,3 @@
 /*! For license information please see index.js.LICENSE.txt */
-!function(e,t){"object"==typeof exports&&"object"==typeof module?module.exports=t(require("react")):"function"==typeof define&&define.amd?define(["react"],t):"object"==typeof exports?exports["token-input"]=t(require("react")):e.TokenInput=t(e.React)}(this,(function(e){return(()=>{var t={733:function(e,t,n){"use strict";var o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var r=o(n(787)),a=o(n(433));t.default=function(){return r.default.createElement("div",{role:"img",className:a.default["delete-button__close-icon"],"aria-hidden":"true"})}},242:function(e,t,n){"use strict";var o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var r=o(n(787)),a=o(n(184)),u=o(n(733)),i=o(n(433)),l=n(754);t.default=function(e){var t=e.onRenderContent,n="function"==typeof t;return r.default.createElement("span",{role:"button",className:(0,a.default)(l.JS__TOKEN__DELETE_BUTTON__CLASS_NAME,i.default["token__delete-button"]),"aria-hidden":"true"},n&&t(),!n&&r.default.createElement(u.default,null))}},83:function(e,t,n){"use strict";var o=this&&this.__createBinding||(Object.create?function(e,t,n,o){void 0===o&&(o=n),Object.defineProperty(e,o,{enumerable:!0,get:function(){return t[n]}})}:function(e,t,n,o){void 0===o&&(o=n),e[o]=t[n]}),r=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),a=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&o(t,e,n);return r(t,e),t},u=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var i=a(n(787)),l=u(n(184)),s=u(n(639)),c=u(n(971)),f=u(n(242)),d=n(754),p=u(n(433)),h=function(e){e.stopPropagation()};t.default=function(e){var t=e.readOnly,n=e.tokenValue,o=e.tokenMeta,r=e.onGetClassName,a=e.onGetDisplayLabel,u=e.onRenderDeleteButtonContent,_=e.onIsEditable,y=e.onGetEditableValue,k=e.onGetErrorMessage,v=e.onBuildTokenValue,T=e.onEditStart,b=e.onEditEnd,E=e.onDelete,O=(0,i.useRef)(null),m=(0,i.useState)(d.DEFAULT_INPUT_INIT_VALUE),g=m[0],C=m[1],N=o.activated,I=o.error,D=(0,i.useMemo)((function(){return _(n,o)}),[_,n,o]),S=(0,i.useCallback)((function(){var e=y(n,o);C(e),T()}),[C,n,o,y,T]);(0,i.useEffect)((function(){var e=O.current;N&&e&&e.getInput().focus()}),[N]);var A=(0,i.useCallback)((function(e){var t=(void 0===e?{}:e).reset,n=void 0!==t&&t,o=0===g.length;if(n||o)b();else{var r=v(g);b(r)}}),[g,v,b]),P=(0,i.useCallback)((function(e){e.stopPropagation(),t||(e.target.closest(".".concat(p.default.token," .").concat(d.JS__TOKEN__DELETE_BUTTON__CLASS_NAME))?E():D&&S())}),[t,D,E,S]),w=(0,i.useCallback)((function(e){var t=e.target.value;C(t)}),[C]),j=(0,i.useCallback)((function(e){(0,c.default)(e,{onEscape:function(){return A({reset:!0})},onEnter:function(){return A()}})}),[A]),V=(0,i.useCallback)((function(){A()}),[A]),M=(0,i.useMemo)((function(){var e;return(0,l.default)(r(n,o),p.default.token,((e={})[p.default["token--read-only"]]=t,e[p.default["token--editable"]]=D&&!t,e[p.default["token--active"]]=N,e[p.default["token--error"]]=I&&!N,e))}),[r,t,D,N,I,n,o]),L=(0,i.useMemo)((function(){return k(n,o)}),[k,n,o]);return N?i.default.createElement("div",{role:"presentation",className:M,onClick:h},i.default.createElement("div",{className:p.default["autosized-wrapper"]},i.default.createElement(s.default,{ref:O,value:g,onChange:w,onKeyDown:j,onBlur:V}))):i.default.createElement("div",{role:"presentation",className:M,onClick:P,title:"string"==typeof L?L:void 0},i.default.createElement("div",{className:p.default["token__label-wrapper"]},a(n,o)),!t&&i.default.createElement(f.default,{onRenderContent:u}))}},84:function(e,t,n){"use strict";var o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var r=o(n(83));t.default=r.default},962:function(e,t,n){"use strict";var o=this&&this.__createBinding||(Object.create?function(e,t,n,o){void 0===o&&(o=n),Object.defineProperty(e,o,{enumerable:!0,get:function(){return t[n]}})}:function(e,t,n,o){void 0===o&&(o=n),e[o]=t[n]}),r=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),a=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&o(t,e,n);return r(t,e),t},u=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var i=a(n(787)),l=u(n(639)),s=u(n(971)),c=u(n(732)),f=n(754),d=u(n(433)),p=(0,i.forwardRef)((function(e,t){var n=e.placeholder,o=e.autoFocus,r=e.onFocus,a=e.onBlur,u=e.separators,p=e.specialKeyDown,h=e.onInputValueChange,_=e.onPreprocess,y=e.onBuildTokenValue,k=e.onNewTokenValuesAppend,v=e.onLastTokenDelete,T=(0,i.useState)(f.DEFAULT_INPUT_INIT_VALUE),b=T[0],E=T[1],O=(0,i.useMemo)((function(){return new RegExp(u.join("|"))}),[u]),m=(0,i.useCallback)((function(e){E(e),h(e,b)}),[h,b,E]),g=(0,i.useCallback)((function(e){if(0!==e.trim().length){var t=e.split(O).filter((function(e){return e.trim().length>0})),n=_(t).map((function(e){return y(e)}));k(n),m(f.DEFAULT_INPUT_INIT_VALUE)}}),[O,_,y,k,m]),C=(0,i.useCallback)((function(e){var t=e.target.value,n=t.substring(t.length-1);!0!==O.test(n)?m(t):g(b)}),[O,g,b,m]),N=(0,c.default)({specialKeyDownConfig:p,inputInitValue:f.DEFAULT_INPUT_INIT_VALUE,inputValue:b,onLastTokenDelete:v,handleInputValueUpdate:m,handleTokensCreate:g}),I=N.handleBackspaceKeyDown,D=N.handleTabKeyDown,S=N.handleEnterKeyDown,A=N.handleEscapeKeyDown,P=(0,i.useCallback)((function(e){(0,s.default)(e,{onBackspace:I,onTab:D,onEnter:S,onEscape:A})}),[I,D,S,A]),w=(0,i.useCallback)((function(){g(b),a()}),[g,b,a]),j=(0,i.useCallback)((function(e){e.preventDefault();var t=e.clipboardData.getData("text");g(t)}),[g]);return i.default.createElement("div",{className:d.default["autosized-wrapper"]},i.default.createElement(l.default,{ref:t,autoFocus:o,placeholder:n,value:b,onChange:C,onKeyDown:P,onPaste:j,onFocus:r,onBlur:w}))}));t.default=p},754:(e,t)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.DEFAULT_INPUT_INIT_VALUE=t.JS__TOKEN__DELETE_BUTTON__CLASS_NAME=t.DEFAULT_SPECIAL_KEY_DOWN_CONFIG=t.KEY_DOWN_HANDLER_CONFIG_OPTION=t.DEFAULT_SEPARATORS=void 0,t.DEFAULT_SEPARATORS=[",",";","\n","\r","\r\n"],t.KEY_DOWN_HANDLER_CONFIG_OPTION={OFF:0,ON:1},t.DEFAULT_SPECIAL_KEY_DOWN_CONFIG={onBackspace:t.KEY_DOWN_HANDLER_CONFIG_OPTION.ON,onTab:t.KEY_DOWN_HANDLER_CONFIG_OPTION.OFF,onEnter:t.KEY_DOWN_HANDLER_CONFIG_OPTION.ON,onEscape:t.KEY_DOWN_HANDLER_CONFIG_OPTION.ON},t.JS__TOKEN__DELETE_BUTTON__CLASS_NAME="js__token__delete-button",t.DEFAULT_INPUT_INIT_VALUE=""},732:function(e,t,n){"use strict";var o=this&&this.__assign||function(){return(o=Object.assign||function(e){for(var t,n=1,o=arguments.length;n<o;n++)for(var r in t=arguments[n])Object.prototype.hasOwnProperty.call(t,r)&&(e[r]=t[r]);return e}).apply(this,arguments)};Object.defineProperty(t,"__esModule",{value:!0});var r=n(787),a=n(754),u=function(e){var t=e.keyDownHandlerConfig,n=e.predefinedHandler;switch(!0){case t===a.KEY_DOWN_HANDLER_CONFIG_OPTION.ON:n()}};t.default=function(e){var t=e.specialKeyDownConfig,n=e.inputInitValue,i=e.inputValue,l=e.onLastTokenDelete,s=e.handleInputValueUpdate,c=e.handleTokensCreate,f=o(o({},a.DEFAULT_SPECIAL_KEY_DOWN_CONFIG),t),d=f.onBackspace,p=f.onTab,h=f.onEnter,_=f.onEscape;return{handleBackspaceKeyDown:(0,r.useCallback)((function(){u({keyDownHandlerConfig:d,predefinedHandler:function(){0===i.length&&l()}})}),[d,i,l]),handleTabKeyDown:(0,r.useCallback)((function(e){u({keyDownHandlerConfig:p,predefinedHandler:function(){e.preventDefault(),c(i)}})}),[p,i,c]),handleEnterKeyDown:(0,r.useCallback)((function(){u({keyDownHandlerConfig:h,predefinedHandler:function(){c(i)}})}),[h,i,c]),handleEscapeKeyDown:(0,r.useCallback)((function(){u({keyDownHandlerConfig:_,predefinedHandler:function(){s(n)}})}),[_,n,s])}}},872:(e,t,n)=>{"use strict";var o=n(787);t.Z=function(){var e=(0,o.useRef)(null),t=(0,o.useCallback)((function(){var t=null==e?void 0:e.current;t&&t.getInput().focus()}),[]);return{tokenCreatorRef:e,focusTokenCreator:t}}},120:function(e,t,n){"use strict";var o=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var o,r=0,a=t.length;r<a;r++)!o&&r in t||(o||(o=Array.prototype.slice.call(t,0,r)),o[r]=t[r]);return e.concat(o||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0});var r=n(787);t.default=function(e){var t=e.tokenValues,n=e.onTokenValuesChange,a=e.focusTokenCreator,u=(0,r.useCallback)((function(e){var r=o([],t,!0);r.splice(e,1),n(r),a()}),[t,n,a]);return{handleTokenDelete:(0,r.useCallback)((function(e){return function(){u(e)}}),[u]),handleLastTokenDelete:(0,r.useCallback)((function(){u(-1)}),[u])}}},970:function(e,t,n){"use strict";var o=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var o,r=0,a=t.length;r<a;r++)!o&&r in t||(o||(o=Array.prototype.slice.call(t,0,r)),o[r]=t[r]);return e.concat(o||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0});var r=n(787);t.default=function(e){var t=e.tokenValues,n=e.onTokenValuesChange,a=e.setTokenActivated,u=e.handleTokenInputFocus,i=e.handleTokenInputBlur;return{handleTokenEditStart:(0,r.useCallback)((function(e){return function(){a(e,!0),u()}}),[a,u]),handleTokenEditEnd:(0,r.useCallback)((function(e){return function(r){if(a(e,!1),i(),void 0!==r){var u=o([],t,!0);u[e]=r,n(u)}}}),[t,n,a,i])}}},598:(e,t,n)=>{"use strict";var o=n(787);t.Z=function(){var e=(0,o.useState)(!1),t=e[0],n=e[1];return{isTokenInputFocused:t,handleTokenInputFocus:(0,o.useCallback)((function(){n(!0)}),[]),handleTokenInputBlur:(0,o.useCallback)((function(){n(!1)}),[])}}},55:function(e,t,n){"use strict";var o=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var o,r=0,a=t.length;r<a;r++)!o&&r in t||(o||(o=Array.prototype.slice.call(t,0,r)),o[r]=t[r]);return e.concat(o||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0});var r=n(787);t.default=function(){var e=(0,r.useState)([]),t=e[0],n=e[1],a=(0,r.useCallback)((function(e,r){if(e>=t.length)throw new Error("setTokenActivated out of tokenMetas scope; targetIndex ".concat(e,"; tokenMetas.length ").concat(t.length));var a=o([],t,!0);a[e].activated=r,n(a)}),[t]);return{tokenMetas:t,setTokenMetas:n,setTokenActivated:a}}},870:function(e,t,n){"use strict";var o=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var o,r=0,a=t.length;r<a;r++)!o&&r in t||(o||(o=Array.prototype.slice.call(t,0,r)),o[r]=t[r]);return e.concat(o||Array.prototype.slice.call(t))},r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=n(787),u=r(n(55)),i=r(n(748));t.default=function(e){var t=e.tokenValues,n=e.onTokenValueValidate,r=(0,u.default)(),l=r.tokenMetas,s=r.setTokenMetas,c=r.setTokenActivated,f=(0,a.useState)(!1),d=f[0],p=f[1],h=(0,a.useState)([]),_=h[0],y=h[1];return(0,a.useLayoutEffect)((function(){var e=!1,r=t.map((function(o,r){var a=n(o,r,t),u=(0,i.default)(a,o,r);return u.error&&!u.activated&&(e=!0),u}));y(o([],t,!0)),s(r),p(e)}),[t,n,s]),{hasInvalidToken:d,internalTokenValues:_,tokenMetas:l,setTokenActivated:c}}},298:(e,t)=>{"use strict";t.Z=function(e){return e.trim()}},748:(e,t)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(e,t,n){return{key:"".concat(JSON.stringify(t),"-").concat(Date.now(),"-").concat(n),activated:!1,error:e}}},309:(e,t)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(){}},300:(e,t)=>{"use strict";t.Z=function(e,t){return"".concat(e)}},707:(e,t)=>{"use strict";t.Z=function(e,t){if("string"==typeof t.error)return t.error}},971:function(e,t,n){"use strict";var o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var r=o(n(776)),a=o(n(309));t.default=function(e,t){var n=t.onBackspace,o=void 0===n?a.default:n,u=t.onTab,i=void 0===u?a.default:u,l=t.onEnter,s=void 0===l?a.default:l,c=t.onEscape,f=void 0===c?a.default:c;switch((0,r.default)(e).key){case"Backspace":o(e);break;case"Tab":i(e);break;case"Enter":s(e);break;case"Escape":f(e)}}},184:(e,t)=>{var n;!function(){"use strict";var o={}.hasOwnProperty;function r(){for(var e=[],t=0;t<arguments.length;t++){var n=arguments[t];if(n){var a=typeof n;if("string"===a||"number"===a)e.push(n);else if(Array.isArray(n)){if(n.length){var u=r.apply(null,n);u&&e.push(u)}}else if("object"===a)if(n.toString===Object.prototype.toString)for(var i in n)o.call(n,i)&&n[i]&&e.push(i);else e.push(n.toString())}}return e.join(" ")}e.exports?(r.default=r,e.exports=r):void 0===(n=function(){return r}.apply(t,[]))||(e.exports=n)}()},776:e=>{self,e.exports=(()=>{"use strict";var e,t={};return e=t,Object.defineProperty(e,"__esModule",{value:!0}),e.default=function(e){var t=e.keyCode,n=e.key;return t&&229===t&&(n="Process"),{key:n}},t})()},433:(e,t,n)=>{"use strict";n.r(t),n.d(t,{default:()=>o});const o={container:"token-input-container","container--focused":"token-input-container--focused","container--errors":"token-input-container--errors","autosized-wrapper":"token-input-autosized-wrapper","token-list":"token-input-token-list",token:"token-input-token","token--error":"token-input-token--error","token--read-only":"token-input-token--read-only","token--editable":"token-input-token--editable","token--active":"token-input-token--active","token__label-wrapper":"token-input-token__label-wrapper","token__delete-button":"token-input-token__delete-button","delete-button__close-icon":"token-input-delete-button__close-icon"}},703:(e,t,n)=>{"use strict";var o=n(414);function r(){}function a(){}a.resetWarningCache=r,e.exports=function(){function e(e,t,n,r,a,u){if(u!==o){var i=new Error("Calling PropTypes validators directly is not supported by the `prop-types` package. Use PropTypes.checkPropTypes() to call them. Read more at http://fb.me/use-check-prop-types");throw i.name="Invariant Violation",i}}function t(){return e}e.isRequired=e;var n={array:e,bigint:e,bool:e,func:e,number:e,object:e,string:e,symbol:e,any:e,arrayOf:t,element:e,elementType:e,instanceOf:t,node:e,objectOf:t,oneOf:t,oneOfType:t,shape:t,exact:t,checkPropTypes:a,resetWarningCache:r};return n.PropTypes=n,n}},697:(e,t,n)=>{e.exports=n(703)()},414:e=>{"use strict";e.exports="SECRET_DO_NOT_PASS_THIS_OR_YOU_WILL_BE_FIRED"},639:(e,t,n)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0});var o=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var o in n)Object.prototype.hasOwnProperty.call(n,o)&&(e[o]=n[o])}return e},r=function(){function e(e,t){for(var n=0;n<t.length;n++){var o=t[n];o.enumerable=o.enumerable||!1,o.configurable=!0,"value"in o&&(o.writable=!0),Object.defineProperty(e,o.key,o)}}return function(t,n,o){return n&&e(t.prototype,n),o&&e(t,o),t}}(),a=n(787),u=l(a),i=l(n(863));function l(e){return e&&e.__esModule?e:{default:e}}var s={position:"absolute",top:0,left:0,visibility:"hidden",height:0,overflow:"scroll",whiteSpace:"pre"},c=["extraWidth","injectStyles","inputClassName","inputRef","inputStyle","minWidth","onAutosize","placeholderIsMinWidth"],f=function(e,t){t.style.fontSize=e.fontSize,t.style.fontFamily=e.fontFamily,t.style.fontWeight=e.fontWeight,t.style.fontStyle=e.fontStyle,t.style.letterSpacing=e.letterSpacing,t.style.textTransform=e.textTransform},d=!("undefined"==typeof window||!window.navigator)&&/MSIE |Trident\/|Edge\//.test(window.navigator.userAgent),p=function(){return d?"_"+Math.random().toString(36).substr(2,12):void 0},h=function(e){function t(e){!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t);var n=function(e,t){if(!e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return!t||"object"!=typeof t&&"function"!=typeof t?e:t}(this,(t.__proto__||Object.getPrototypeOf(t)).call(this,e));return n.inputRef=function(e){n.input=e,"function"==typeof n.props.inputRef&&n.props.inputRef(e)},n.placeHolderSizerRef=function(e){n.placeHolderSizer=e},n.sizerRef=function(e){n.sizer=e},n.state={inputWidth:e.minWidth,inputId:e.id||p(),prevId:e.id},n}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function, not "+typeof t);e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,enumerable:!1,writable:!0,configurable:!0}}),t&&(Object.setPrototypeOf?Object.setPrototypeOf(e,t):e.__proto__=t)}(t,e),r(t,null,[{key:"getDerivedStateFromProps",value:function(e,t){var n=e.id;return n!==t.prevId?{inputId:n||p(),prevId:n}:null}}]),r(t,[{key:"componentDidMount",value:function(){this.mounted=!0,this.copyInputStyles(),this.updateInputWidth()}},{key:"componentDidUpdate",value:function(e,t){t.inputWidth!==this.state.inputWidth&&"function"==typeof this.props.onAutosize&&this.props.onAutosize(this.state.inputWidth),this.updateInputWidth()}},{key:"componentWillUnmount",value:function(){this.mounted=!1}},{key:"copyInputStyles",value:function(){if(this.mounted&&window.getComputedStyle){var e=this.input&&window.getComputedStyle(this.input);e&&(f(e,this.sizer),this.placeHolderSizer&&f(e,this.placeHolderSizer))}}},{key:"updateInputWidth",value:function(){if(this.mounted&&this.sizer&&void 0!==this.sizer.scrollWidth){var e=void 0;e=this.props.placeholder&&(!this.props.value||this.props.value&&this.props.placeholderIsMinWidth)?Math.max(this.sizer.scrollWidth,this.placeHolderSizer.scrollWidth)+2:this.sizer.scrollWidth+2,(e+="number"===this.props.type&&void 0===this.props.extraWidth?16:parseInt(this.props.extraWidth)||0)<this.props.minWidth&&(e=this.props.minWidth),e!==this.state.inputWidth&&this.setState({inputWidth:e})}}},{key:"getInput",value:function(){return this.input}},{key:"focus",value:function(){this.input.focus()}},{key:"blur",value:function(){this.input.blur()}},{key:"select",value:function(){this.input.select()}},{key:"renderStyles",value:function(){var e=this.props.injectStyles;return d&&e?u.default.createElement("style",{dangerouslySetInnerHTML:{__html:"input#"+this.state.inputId+"::-ms-clear {display: none;}"}}):null}},{key:"render",value:function(){var e=[this.props.defaultValue,this.props.value,""].reduce((function(e,t){return null!=e?e:t})),t=o({},this.props.style);t.display||(t.display="inline-block");var n=o({boxSizing:"content-box",width:this.state.inputWidth+"px"},this.props.inputStyle),r=function(e,t){var n={};for(var o in e)t.indexOf(o)>=0||Object.prototype.hasOwnProperty.call(e,o)&&(n[o]=e[o]);return n}(this.props,[]);return function(e){c.forEach((function(t){return delete e[t]}))}(r),r.className=this.props.inputClassName,r.id=this.state.inputId,r.style=n,u.default.createElement("div",{className:this.props.className,style:t},this.renderStyles(),u.default.createElement("input",o({},r,{ref:this.inputRef})),u.default.createElement("div",{ref:this.sizerRef,style:s},e),this.props.placeholder?u.default.createElement("div",{ref:this.placeHolderSizerRef,style:s},this.props.placeholder):null)}}]),t}(a.Component);h.propTypes={className:i.default.string,defaultValue:i.default.any,extraWidth:i.default.oneOfType([i.default.number,i.default.string]),id:i.default.string,injectStyles:i.default.bool,inputClassName:i.default.string,inputRef:i.default.func,inputStyle:i.default.object,minWidth:i.default.oneOfType([i.default.number,i.default.string]),onAutosize:i.default.func,onChange:i.default.func,placeholder:i.default.string,placeholderIsMinWidth:i.default.bool,style:i.default.object,value:i.default.any},h.defaultProps={minWidth:1,injectStyles:!0},t.default=h},216:(e,t,n)=>{"use strict";var o=n(138);function r(){}function a(){}a.resetWarningCache=r,e.exports=function(){function e(e,t,n,r,a,u){if(u!==o){var i=new Error("Calling PropTypes validators directly is not supported by the `prop-types` package. Use PropTypes.checkPropTypes() to call them. Read more at http://fb.me/use-check-prop-types");throw i.name="Invariant Violation",i}}function t(){return e}e.isRequired=e;var n={array:e,bool:e,func:e,number:e,object:e,string:e,symbol:e,any:e,arrayOf:t,element:e,elementType:e,instanceOf:t,node:e,objectOf:t,oneOf:t,oneOfType:t,shape:t,exact:t,checkPropTypes:a,resetWarningCache:r};return n.PropTypes=n,n}},863:(e,t,n)=>{e.exports=n(216)()},138:e=>{"use strict";e.exports="SECRET_DO_NOT_PASS_THIS_OR_YOU_WILL_BE_FIRED"},787:t=>{"use strict";t.exports=e}},n={};function o(e){var r=n[e];if(void 0!==r)return r.exports;var a=n[e]={exports:{}};return t[e].call(a.exports,a,a.exports,o),a.exports}o.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return o.d(t,{a:t}),t},o.d=(e,t)=>{for(var n in t)o.o(t,n)&&!o.o(e,n)&&Object.defineProperty(e,n,{enumerable:!0,get:t[n]})},o.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t),o.r=e=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})};var r={};return(()=>{"use strict";o.r(r),o.d(r,{JS__TOKEN__DELETE_BUTTON__CLASS_NAME:()=>g.JS__TOKEN__DELETE_BUTTON__CLASS_NAME,KEY_DOWN_HANDLER_CONFIG_OPTION:()=>g.KEY_DOWN_HANDLER_CONFIG_OPTION,default:()=>w});var e=o(787),t=o.n(e),n=o(697),a=o.n(n),u=o(184),i=o.n(u),l=o(962),s=o.n(l),c=o(84),f=o.n(c),d=o(598),p=o(872),h=o(870),_=o.n(h),y=o(970),k=o.n(y),v=o(120),T=o.n(v),b=o(309),E=o(298),O=o(300),m=o(707),g=o(754),C=o(433),N=["className","readOnly","autoFocus","placeholder","tokenValues","onTokenValuesChange","onBuildTokenValue","customizeTokenComponent","onGetTokenClassName","onGetTokenDisplayLabel","onRenderTokenDeleteButtonContent","onIsTokenEditable","onGetTokenEditableValue","onGetTokenErrorMessage","separators","specialKeyDown","onPreprocess","onInputValueChange","onTokenValueValidate"];function I(){return(I=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var o in n)Object.prototype.hasOwnProperty.call(n,o)&&(e[o]=n[o])}return e}).apply(this,arguments)}function D(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}function S(e){return function(e){if(Array.isArray(e))return A(e)}(e)||function(e){if("undefined"!=typeof Symbol&&null!=e[Symbol.iterator]||null!=e["@@iterator"])return Array.from(e)}(e)||function(e,t){if(e){if("string"==typeof e)return A(e,t);var n=Object.prototype.toString.call(e).slice(8,-1);return"Object"===n&&e.constructor&&(n=e.constructor.name),"Map"===n||"Set"===n?Array.from(e):"Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)?A(e,t):void 0}}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function A(e,t){(null==t||t>e.length)&&(t=e.length);for(var n=0,o=new Array(t);n<t;n++)o[n]=e[n];return o}var P=function(n){var o,r=n.className,a=n.readOnly,u=n.autoFocus,l=n.placeholder,c=n.tokenValues,h=n.onTokenValuesChange,y=n.onBuildTokenValue,v=n.customizeTokenComponent,b=n.onGetTokenClassName,E=n.onGetTokenDisplayLabel,O=n.onRenderTokenDeleteButtonContent,m=n.onIsTokenEditable,g=n.onGetTokenEditableValue,A=n.onGetTokenErrorMessage,P=n.separators,w=n.specialKeyDown,j=n.onPreprocess,V=n.onInputValueChange,M=n.onTokenValueValidate,L=function(e,t){if(null==e)return{};var n,o,r=function(e,t){if(null==e)return{};var n,o,r={},a=Object.keys(e);for(o=0;o<a.length;o++)n=a[o],t.indexOf(n)>=0||(r[n]=e[n]);return r}(e,t);if(Object.getOwnPropertySymbols){var a=Object.getOwnPropertySymbols(e);for(o=0;o<a.length;o++)n=a[o],t.indexOf(n)>=0||Object.prototype.propertyIsEnumerable.call(e,n)&&(r[n]=e[n])}return r}(n,N),R=(0,d.Z)(),F=R.isTokenInputFocused,W=R.handleTokenInputFocus,x=R.handleTokenInputBlur,B=(0,p.Z)(),U=B.tokenCreatorRef,G=B.focusTokenCreator,K=_()({tokenValues:c,onTokenValueValidate:M}),z=K.hasInvalidToken,H=K.internalTokenValues,Y=K.tokenMetas,Z=K.setTokenActivated,J=k()({tokenValues:c,onTokenValuesChange:h,setTokenActivated:Z,handleTokenInputFocus:W,handleTokenInputBlur:x}),q=J.handleTokenEditStart,$=J.handleTokenEditEnd,Q=T()({tokenValues:c,onTokenValuesChange:h,focusTokenCreator:G}),X=Q.handleTokenDelete,ee=Q.handleLastTokenDelete,te=(0,e.useCallback)((function(e){if(0!==e.length){var t=[].concat(S(c),S(e));h(t)}}),[c,h]),ne=(0,e.useCallback)((function(e){if("function"!=typeof j)return e;var t=j(e);if(!0!==Array.isArray(t))throw new Error("onPreprocess should return an array of values");return t}),[j]),oe=v||f();return t().createElement("div",I({},L,{className:i()(r,C.default.container,(o={},D(o,C.default["container--focused"],F),D(o,C.default["container--errors"],z),o)),onClick:G,role:"presentation","data-component-name":"TokenInput"}),t().createElement("div",{className:C.default["token-list"]},H.map((function(e,n){var o=Y[n],r=o.key;return t().createElement(oe,{key:r,readOnly:a,tokenValue:e,tokenMeta:o,onGetClassName:b,onGetDisplayLabel:E,onRenderDeleteButtonContent:O,onIsEditable:m,onGetEditableValue:g,onGetErrorMessage:A,onBuildTokenValue:y,onEditStart:q(n),onEditEnd:$(n),onDelete:X(n)})}))),!a&&t().createElement(s(),{ref:U,placeholder:l,autoFocus:u,onFocus:W,onBlur:x,separators:P,specialKeyDown:w,onInputValueChange:V,onPreprocess:ne,onBuildTokenValue:y,onNewTokenValuesAppend:te,onLastTokenDelete:ee}))};P.propTypes={style:a().object,className:a().string,readOnly:a().bool,autoFocus:a().bool,placeholder:a().string,separators:a().array,tokenValues:a().array.isRequired,onTokenValuesChange:a().func,onPreprocess:a().func,onBuildTokenValue:a().func,onInputValueChange:a().func,onTokenValueValidate:a().func,onGetTokenClassName:a().func,onGetTokenDisplayLabel:a().func,onRenderTokenDeleteButtonContent:a().func,onIsTokenEditable:a().func,onGetTokenEditableValue:a().func,onGetTokenErrorMessage:a().func,customizeTokenComponent:a().func,specialKeyDown:a().object},P.defaultProps={className:"",readOnly:!1,autoFocus:!1,placeholder:"",separators:g.DEFAULT_SEPARATORS,specialKeyDown:{onBackspace:1,onTab:0,onEnter:1,onEscape:1},onBuildTokenValue:E.Z,onInputValueChange:b.default,onTokenValueValidate:b.default,onTokenValuesChange:b.default,onGetTokenClassName:b.default,onGetTokenDisplayLabel:O.Z,onIsTokenEditable:function(){return!0},onGetTokenEditableValue:O.Z,onGetTokenErrorMessage:m.Z};const w=P})(),r})()}));
+!function(e,t){"object"==typeof exports&&"object"==typeof module?module.exports=t(require("react")):"function"==typeof define&&define.amd?define(["react"],t):"object"==typeof exports?exports["token-input"]=t(require("react")):e.TokenInput=t(e.React)}(this,(function(e){return(()=>{var t={733:function(e,t,n){"use strict";var o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var r=o(n(787)),a=o(n(433));t.default=function(){return r.default.createElement("div",{role:"img",className:a.default["delete-button__close-icon"],"aria-hidden":"true"})}},242:function(e,t,n){"use strict";var o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var r=o(n(787)),a=o(n(184)),u=o(n(733)),i=o(n(433)),l=n(754);t.default=function(e){var t=e.onRenderContent,n="function"==typeof t;return r.default.createElement("span",{role:"button",className:(0,a.default)(l.JS__TOKEN__DELETE_BUTTON__CLASS_NAME,i.default["token__delete-button"]),"aria-hidden":"true"},n&&t(),!n&&r.default.createElement(u.default,null))}},83:function(e,t,n){"use strict";var o=this&&this.__createBinding||(Object.create?function(e,t,n,o){void 0===o&&(o=n),Object.defineProperty(e,o,{enumerable:!0,get:function(){return t[n]}})}:function(e,t,n,o){void 0===o&&(o=n),e[o]=t[n]}),r=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),a=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&o(t,e,n);return r(t,e),t},u=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var i=a(n(787)),l=u(n(184)),s=u(n(639)),c=u(n(971)),d=u(n(242)),f=n(754),p=u(n(433)),_=function(e){e.stopPropagation()};t.default=function(e){var t=e.readOnly,n=e.tokenValue,o=e.tokenMeta,r=e.onGetClassName,a=e.onGetDisplayLabel,u=e.onRenderDeleteButtonContent,h=e.onIsEditable,v=e.onGetEditableValue,y=e.onGetErrorMessage,k=e.onBuildTokenValue,b=e.onEditStart,T=e.onEditEnd,O=e.onDelete,E=(0,i.useRef)(null),m=(0,i.useState)(f.DEFAULT_INPUT_INIT_VALUE),g=m[0],C=m[1],N=o.activated,D=o.error,I=(0,i.useMemo)((function(){var e;return null===(e=null==h?void 0:h(n,o))||void 0===e||e}),[h,n,o]),S=(0,i.useCallback)((function(){var e=v(n,o);C(e),b()}),[C,n,o,v,b]);(0,i.useEffect)((function(){var e=E.current;N&&e&&e.getInput().focus()}),[N]);var P=(0,i.useCallback)((function(e){var t=(void 0===e?{}:e).reset,n=void 0!==t&&t,o=0===g.length;if(n||o)T();else{var r=k(g);T(r)}}),[g,k,T]),A=(0,i.useCallback)((function(e){e.stopPropagation(),t||(e.target.closest(".".concat(p.default.token," .").concat(f.JS__TOKEN__DELETE_BUTTON__CLASS_NAME))?O():I&&S())}),[t,I,O,S]),j=(0,i.useCallback)((function(e){var t=e.target.value;C(t)}),[C]),w=(0,i.useCallback)((function(e){(0,c.default)(e,{onEscape:function(){return P({reset:!0})},onEnter:function(){return P()}})}),[P]),M=(0,i.useCallback)((function(){P()}),[P]),L=(0,i.useMemo)((function(){var e;return(0,l.default)(r(n,o),p.default.token,((e={})[p.default["token--read-only"]]=t,e[p.default["token--editable"]]=I&&!t,e[p.default["token--active"]]=N,e[p.default["token--error"]]=D&&!N,e))}),[r,t,I,N,D,n,o]),V=(0,i.useMemo)((function(){return y(n,o)}),[y,n,o]);return N?i.default.createElement("div",{role:"presentation",className:L,onClick:_},i.default.createElement("div",{className:p.default["autosized-wrapper"]},i.default.createElement(s.default,{ref:E,value:g,onChange:j,onKeyDown:w,onBlur:M}))):i.default.createElement("div",{role:"presentation",className:L,onClick:A,title:"string"==typeof V?V:void 0},i.default.createElement("div",{className:p.default["token__label-wrapper"]},a(n,o)),!t&&i.default.createElement(d.default,{onRenderContent:u}))}},84:function(e,t,n){"use strict";var o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var r=o(n(83));t.default=r.default},962:function(e,t,n){"use strict";var o=this&&this.__createBinding||(Object.create?function(e,t,n,o){void 0===o&&(o=n),Object.defineProperty(e,o,{enumerable:!0,get:function(){return t[n]}})}:function(e,t,n,o){void 0===o&&(o=n),e[o]=t[n]}),r=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),a=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&o(t,e,n);return r(t,e),t},u=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var i=a(n(787)),l=u(n(639)),s=u(n(971)),c=u(n(732)),d=n(754),f=u(n(433)),p=(0,i.forwardRef)((function(e,t){var n=e.placeholder,o=e.autoFocus,r=e.onFocus,a=e.onBlur,u=e.separators,p=e.specialKeyDown,_=e.onInputValueChange,h=e.onPreprocess,v=e.onBuildTokenValue,y=e.onNewTokenValuesAppend,k=e.onLastTokenDelete,b=(0,i.useState)(d.DEFAULT_INPUT_INIT_VALUE),T=b[0],O=b[1],E=(0,i.useMemo)((function(){return new RegExp(u.join("|"))}),[u]),m=(0,i.useCallback)((function(e){O(e),_(e,T)}),[_,T,O]),g=(0,i.useCallback)((function(e){if(0!==e.trim().length){var t=e.split(E).filter((function(e){return e.trim().length>0})),n=h(t).map((function(e){return v(e)}));y(n),m(d.DEFAULT_INPUT_INIT_VALUE)}}),[E,h,v,y,m]),C=(0,i.useCallback)((function(e){var t=e.target.value,n=t.substring(t.length-1);!0!==E.test(n)?m(t):g(T)}),[E,g,T,m]),N=(0,c.default)({specialKeyDownConfig:p,inputInitValue:d.DEFAULT_INPUT_INIT_VALUE,inputValue:T,onLastTokenDelete:k,handleInputValueUpdate:m,handleTokensCreate:g}),D=N.handleBackspaceKeyDown,I=N.handleTabKeyDown,S=N.handleEnterKeyDown,P=N.handleEscapeKeyDown,A=(0,i.useCallback)((function(e){(0,s.default)(e,{onBackspace:D,onTab:I,onEnter:S,onEscape:P})}),[D,I,S,P]),j=(0,i.useCallback)((function(){g(T),a()}),[g,T,a]),w=(0,i.useCallback)((function(e){e.preventDefault();var t=e.clipboardData.getData("text");g(t)}),[g]);return i.default.createElement("div",{className:f.default["autosized-wrapper"]},i.default.createElement(l.default,{ref:t,autoFocus:o,placeholder:n,value:T,onChange:C,onKeyDown:A,onPaste:w,onFocus:r,onBlur:j}))}));t.default=p},922:function(e,t,n){"use strict";var o=this&&this.__assign||function(){return(o=Object.assign||function(e){for(var t,n=1,o=arguments.length;n<o;n++)for(var r in t=arguments[n])Object.prototype.hasOwnProperty.call(t,r)&&(e[r]=t[r]);return e}).apply(this,arguments)},r=this&&this.__createBinding||(Object.create?function(e,t,n,o){void 0===o&&(o=n),Object.defineProperty(e,o,{enumerable:!0,get:function(){return t[n]}})}:function(e,t,n,o){void 0===o&&(o=n),e[o]=t[n]}),a=this&&this.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),u=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(null!=e)for(var n in e)"default"!==n&&Object.prototype.hasOwnProperty.call(e,n)&&r(t,e,n);return a(t,e),t},i=this&&this.__rest||function(e,t){var n={};for(var o in e)Object.prototype.hasOwnProperty.call(e,o)&&t.indexOf(o)<0&&(n[o]=e[o]);if(null!=e&&"function"==typeof Object.getOwnPropertySymbols){var r=0;for(o=Object.getOwnPropertySymbols(e);r<o.length;r++)t.indexOf(o[r])<0&&Object.prototype.propertyIsEnumerable.call(e,o[r])&&(n[o[r]]=e[o[r]])}return n},l=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var o,r=0,a=t.length;r<a;r++)!o&&r in t||(o||(o=Array.prototype.slice.call(t,0,r)),o[r]=t[r]);return e.concat(o||Array.prototype.slice.call(t))},s=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var c=u(n(787)),d=s(n(184)),f=s(n(962)),p=s(n(84)),_=s(n(598)),h=s(n(872)),v=s(n(870)),y=s(n(970)),k=s(n(120)),b=s(n(309)),T=s(n(298)),O=s(n(300)),E=s(n(707)),m=n(754),g=s(n(433));t.default=function(e){var t,n=e.className,r=void 0===n?"":n,a=e.readOnly,u=void 0!==a&&a,s=e.autoFocus,C=void 0!==s&&s,N=e.placeholder,D=void 0===N?"":N,I=e.tokenValues,S=e.onBuildTokenValue,P=void 0===S?T.default:S,A=e.onTokenValuesChange,j=void 0===A?b.default:A,w=e.customizeTokenComponent,M=e.onGetTokenClassName,L=void 0===M?b.default:M,V=e.onGetTokenDisplayLabel,F=void 0===V?O.default:V,W=e.onRenderTokenDeleteButtonContent,R=e.onIsTokenEditable,U=e.onGetTokenEditableValue,x=void 0===U?O.default:U,B=e.onGetTokenErrorMessage,K=void 0===B?E.default:B,G=e.separators,z=void 0===G?m.DEFAULT_SEPARATORS:G,H=e.specialKeyDown,Y=void 0===H?m.DEFAULT_SPECIAL_KEY_DOWN_CONFIG:H,J=e.onPreprocess,q=e.onInputValueChange,Q=void 0===q?b.default:q,X=e.onTokenValueValidate,Z=void 0===X?b.default:X,$=i(e,["className","readOnly","autoFocus","placeholder","tokenValues","onBuildTokenValue","onTokenValuesChange","customizeTokenComponent","onGetTokenClassName","onGetTokenDisplayLabel","onRenderTokenDeleteButtonContent","onIsTokenEditable","onGetTokenEditableValue","onGetTokenErrorMessage","separators","specialKeyDown","onPreprocess","onInputValueChange","onTokenValueValidate"]),ee=(0,_.default)(),te=ee.isTokenInputFocused,ne=ee.handleTokenInputFocus,oe=ee.handleTokenInputBlur,re=(0,h.default)(),ae=re.tokenCreatorRef,ue=re.focusTokenCreator,ie=(0,v.default)({tokenValues:I,onTokenValueValidate:Z}),le=ie.hasInvalidToken,se=ie.internalTokenValues,ce=ie.tokenMetas,de=ie.setTokenActivated,fe=(0,y.default)({tokenValues:I,onTokenValuesChange:j,setTokenActivated:de,handleTokenInputFocus:ne,handleTokenInputBlur:oe}),pe=fe.handleTokenEditStart,_e=fe.handleTokenEditEnd,he=(0,k.default)({tokenValues:I,onTokenValuesChange:j,focusTokenCreator:ue}),ve=he.handleTokenDelete,ye=he.handleLastTokenDelete,ke=(0,c.useCallback)((function(e){if(0!==e.length){var t=l(l([],I,!0),e,!0);j(t)}}),[I,j]),be=(0,c.useCallback)((function(e){if("function"!=typeof J)return e;var t=J(e);if(!0!==Array.isArray(t))throw new Error("onPreprocess should return an array of values");return t}),[J]),Te=w||p.default;return c.default.createElement("div",o({},$,{className:(0,d.default)(r,g.default.container,(t={},t[g.default["container--focused"]]=te,t[g.default["container--errors"]]=le,t)),onClick:ue,role:"presentation"}),c.default.createElement("div",{className:g.default["token-list"]},se.map((function(e,t){var n=ce[t],o=n.key;return c.default.createElement(Te,{key:o,readOnly:u,tokenValue:e,tokenMeta:n,onGetClassName:L,onGetDisplayLabel:F,onRenderDeleteButtonContent:W,onIsEditable:R,onGetEditableValue:x,onGetErrorMessage:K,onBuildTokenValue:P,onEditStart:pe(t),onEditEnd:_e(t),onDelete:ve(t)})}))),!u&&c.default.createElement(f.default,{ref:ae,placeholder:D,autoFocus:C,onFocus:ne,onBlur:oe,separators:z,specialKeyDown:Y,onInputValueChange:Q,onPreprocess:be,onBuildTokenValue:P,onNewTokenValuesAppend:ke,onLastTokenDelete:ye}))}},754:(e,t)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.DEFAULT_INPUT_INIT_VALUE=t.JS__TOKEN__DELETE_BUTTON__CLASS_NAME=t.DEFAULT_SPECIAL_KEY_DOWN_CONFIG=t.KEY_DOWN_HANDLER_CONFIG_OPTION=t.DEFAULT_SEPARATORS=void 0,t.DEFAULT_SEPARATORS=[",",";","\n","\r","\r\n"],t.KEY_DOWN_HANDLER_CONFIG_OPTION={OFF:0,ON:1},t.DEFAULT_SPECIAL_KEY_DOWN_CONFIG={onBackspace:t.KEY_DOWN_HANDLER_CONFIG_OPTION.ON,onTab:t.KEY_DOWN_HANDLER_CONFIG_OPTION.OFF,onEnter:t.KEY_DOWN_HANDLER_CONFIG_OPTION.ON,onEscape:t.KEY_DOWN_HANDLER_CONFIG_OPTION.ON},t.JS__TOKEN__DELETE_BUTTON__CLASS_NAME="js__token__delete-button",t.DEFAULT_INPUT_INIT_VALUE=""},732:function(e,t,n){"use strict";var o=this&&this.__assign||function(){return(o=Object.assign||function(e){for(var t,n=1,o=arguments.length;n<o;n++)for(var r in t=arguments[n])Object.prototype.hasOwnProperty.call(t,r)&&(e[r]=t[r]);return e}).apply(this,arguments)};Object.defineProperty(t,"__esModule",{value:!0});var r=n(787),a=n(754),u=function(e){var t=e.keyDownHandlerConfig,n=e.predefinedHandler;switch(!0){case t===a.KEY_DOWN_HANDLER_CONFIG_OPTION.ON:n()}};t.default=function(e){var t=e.specialKeyDownConfig,n=e.inputInitValue,i=e.inputValue,l=e.onLastTokenDelete,s=e.handleInputValueUpdate,c=e.handleTokensCreate,d=o(o({},a.DEFAULT_SPECIAL_KEY_DOWN_CONFIG),t),f=d.onBackspace,p=d.onTab,_=d.onEnter,h=d.onEscape;return{handleBackspaceKeyDown:(0,r.useCallback)((function(){u({keyDownHandlerConfig:f,predefinedHandler:function(){0===i.length&&l()}})}),[f,i,l]),handleTabKeyDown:(0,r.useCallback)((function(e){u({keyDownHandlerConfig:p,predefinedHandler:function(){e.preventDefault(),c(i)}})}),[p,i,c]),handleEnterKeyDown:(0,r.useCallback)((function(){u({keyDownHandlerConfig:_,predefinedHandler:function(){c(i)}})}),[_,i,c]),handleEscapeKeyDown:(0,r.useCallback)((function(){u({keyDownHandlerConfig:h,predefinedHandler:function(){s(n)}})}),[h,n,s])}}},872:(e,t,n)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0});var o=n(787);t.default=function(){var e=(0,o.useRef)(null),t=(0,o.useCallback)((function(){var t=null==e?void 0:e.current;t&&t.getInput().focus()}),[]);return{tokenCreatorRef:e,focusTokenCreator:t}}},120:function(e,t,n){"use strict";var o=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var o,r=0,a=t.length;r<a;r++)!o&&r in t||(o||(o=Array.prototype.slice.call(t,0,r)),o[r]=t[r]);return e.concat(o||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0});var r=n(787);t.default=function(e){var t=e.tokenValues,n=e.onTokenValuesChange,a=e.focusTokenCreator,u=(0,r.useCallback)((function(e){var r=o([],t,!0);r.splice(e,1),n(r),a()}),[t,n,a]);return{handleTokenDelete:(0,r.useCallback)((function(e){return function(){u(e)}}),[u]),handleLastTokenDelete:(0,r.useCallback)((function(){u(-1)}),[u])}}},970:function(e,t,n){"use strict";var o=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var o,r=0,a=t.length;r<a;r++)!o&&r in t||(o||(o=Array.prototype.slice.call(t,0,r)),o[r]=t[r]);return e.concat(o||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0});var r=n(787);t.default=function(e){var t=e.tokenValues,n=e.onTokenValuesChange,a=e.setTokenActivated,u=e.handleTokenInputFocus,i=e.handleTokenInputBlur;return{handleTokenEditStart:(0,r.useCallback)((function(e){return function(){a(e,!0),u()}}),[a,u]),handleTokenEditEnd:(0,r.useCallback)((function(e){return function(r){if(a(e,!1),i(),void 0!==r){var u=o([],t,!0);u[e]=r,n(u)}}}),[t,n,a,i])}}},598:(e,t,n)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0});var o=n(787);t.default=function(){var e=(0,o.useState)(!1),t=e[0],n=e[1];return{isTokenInputFocused:t,handleTokenInputFocus:(0,o.useCallback)((function(){n(!0)}),[]),handleTokenInputBlur:(0,o.useCallback)((function(){n(!1)}),[])}}},55:function(e,t,n){"use strict";var o=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var o,r=0,a=t.length;r<a;r++)!o&&r in t||(o||(o=Array.prototype.slice.call(t,0,r)),o[r]=t[r]);return e.concat(o||Array.prototype.slice.call(t))};Object.defineProperty(t,"__esModule",{value:!0});var r=n(787);t.default=function(){var e=(0,r.useState)([]),t=e[0],n=e[1],a=(0,r.useCallback)((function(e,r){if(e>=t.length)throw new Error("setTokenActivated out of tokenMetas scope; targetIndex ".concat(e,"; tokenMetas.length ").concat(t.length));var a=o([],t,!0);a[e].activated=r,n(a)}),[t]);return{tokenMetas:t,setTokenMetas:n,setTokenActivated:a}}},870:function(e,t,n){"use strict";var o=this&&this.__spreadArray||function(e,t,n){if(n||2===arguments.length)for(var o,r=0,a=t.length;r<a;r++)!o&&r in t||(o||(o=Array.prototype.slice.call(t,0,r)),o[r]=t[r]);return e.concat(o||Array.prototype.slice.call(t))},r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var a=n(787),u=r(n(55)),i=r(n(748));t.default=function(e){var t=e.tokenValues,n=e.onTokenValueValidate,r=(0,u.default)(),l=r.tokenMetas,s=r.setTokenMetas,c=r.setTokenActivated,d=(0,a.useState)(!1),f=d[0],p=d[1],_=(0,a.useState)([]),h=_[0],v=_[1];return(0,a.useLayoutEffect)((function(){var e=!1,r=t.map((function(o,r){var a=n(o,r,t),u=(0,i.default)(a,o,r);return u.error&&!u.activated&&(e=!0),u}));v(o([],t,!0)),s(r),p(e)}),[t,n,s]),{hasInvalidToken:f,internalTokenValues:h,tokenMetas:l,setTokenActivated:c}}},298:(e,t)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(e){return e.trim()}},748:(e,t)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(e,t,n){return{key:"".concat(JSON.stringify(t),"-").concat(Date.now(),"-").concat(n),activated:!1,error:e}}},309:(e,t)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(){}},300:(e,t)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(e,t){return"".concat(e)}},707:(e,t)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.default=function(e,t){if("string"==typeof t.error)return t.error}},971:function(e,t,n){"use strict";var o=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0});var r=o(n(776)),a=o(n(309));t.default=function(e,t){var n=t.onBackspace,o=void 0===n?a.default:n,u=t.onTab,i=void 0===u?a.default:u,l=t.onEnter,s=void 0===l?a.default:l,c=t.onEscape,d=void 0===c?a.default:c;switch((0,r.default)(e).key){case"Backspace":o(e);break;case"Tab":i(e);break;case"Enter":s(e);break;case"Escape":d(e)}}},184:(e,t)=>{var n;!function(){"use strict";var o={}.hasOwnProperty;function r(){for(var e=[],t=0;t<arguments.length;t++){var n=arguments[t];if(n){var a=typeof n;if("string"===a||"number"===a)e.push(n);else if(Array.isArray(n)){if(n.length){var u=r.apply(null,n);u&&e.push(u)}}else if("object"===a)if(n.toString===Object.prototype.toString)for(var i in n)o.call(n,i)&&n[i]&&e.push(i);else e.push(n.toString())}}return e.join(" ")}e.exports?(r.default=r,e.exports=r):void 0===(n=function(){return r}.apply(t,[]))||(e.exports=n)}()},776:e=>{self,e.exports=(()=>{"use strict";var e,t={};return e=t,Object.defineProperty(e,"__esModule",{value:!0}),e.default=function(e){var t=e.keyCode,n=e.key;return t&&229===t&&(n="Process"),{key:n}},t})()},433:(e,t,n)=>{"use strict";n.r(t),n.d(t,{default:()=>o});const o={container:"token-input-container","container--focused":"token-input-container--focused","container--errors":"token-input-container--errors","autosized-wrapper":"token-input-autosized-wrapper","token-list":"token-input-token-list",token:"token-input-token","token--error":"token-input-token--error","token--read-only":"token-input-token--read-only","token--editable":"token-input-token--editable","token--active":"token-input-token--active","token__label-wrapper":"token-input-token__label-wrapper","token__delete-button":"token-input-token__delete-button","delete-button__close-icon":"token-input-delete-button__close-icon"}},639:(e,t,n)=>{"use strict";Object.defineProperty(t,"__esModule",{value:!0});var o=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var o in n)Object.prototype.hasOwnProperty.call(n,o)&&(e[o]=n[o])}return e},r=function(){function e(e,t){for(var n=0;n<t.length;n++){var o=t[n];o.enumerable=o.enumerable||!1,o.configurable=!0,"value"in o&&(o.writable=!0),Object.defineProperty(e,o.key,o)}}return function(t,n,o){return n&&e(t.prototype,n),o&&e(t,o),t}}(),a=n(787),u=l(a),i=l(n(863));function l(e){return e&&e.__esModule?e:{default:e}}var s={position:"absolute",top:0,left:0,visibility:"hidden",height:0,overflow:"scroll",whiteSpace:"pre"},c=["extraWidth","injectStyles","inputClassName","inputRef","inputStyle","minWidth","onAutosize","placeholderIsMinWidth"],d=function(e,t){t.style.fontSize=e.fontSize,t.style.fontFamily=e.fontFamily,t.style.fontWeight=e.fontWeight,t.style.fontStyle=e.fontStyle,t.style.letterSpacing=e.letterSpacing,t.style.textTransform=e.textTransform},f=!("undefined"==typeof window||!window.navigator)&&/MSIE |Trident\/|Edge\//.test(window.navigator.userAgent),p=function(){return f?"_"+Math.random().toString(36).substr(2,12):void 0},_=function(e){function t(e){!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t);var n=function(e,t){if(!e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return!t||"object"!=typeof t&&"function"!=typeof t?e:t}(this,(t.__proto__||Object.getPrototypeOf(t)).call(this,e));return n.inputRef=function(e){n.input=e,"function"==typeof n.props.inputRef&&n.props.inputRef(e)},n.placeHolderSizerRef=function(e){n.placeHolderSizer=e},n.sizerRef=function(e){n.sizer=e},n.state={inputWidth:e.minWidth,inputId:e.id||p(),prevId:e.id},n}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function, not "+typeof t);e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,enumerable:!1,writable:!0,configurable:!0}}),t&&(Object.setPrototypeOf?Object.setPrototypeOf(e,t):e.__proto__=t)}(t,e),r(t,null,[{key:"getDerivedStateFromProps",value:function(e,t){var n=e.id;return n!==t.prevId?{inputId:n||p(),prevId:n}:null}}]),r(t,[{key:"componentDidMount",value:function(){this.mounted=!0,this.copyInputStyles(),this.updateInputWidth()}},{key:"componentDidUpdate",value:function(e,t){t.inputWidth!==this.state.inputWidth&&"function"==typeof this.props.onAutosize&&this.props.onAutosize(this.state.inputWidth),this.updateInputWidth()}},{key:"componentWillUnmount",value:function(){this.mounted=!1}},{key:"copyInputStyles",value:function(){if(this.mounted&&window.getComputedStyle){var e=this.input&&window.getComputedStyle(this.input);e&&(d(e,this.sizer),this.placeHolderSizer&&d(e,this.placeHolderSizer))}}},{key:"updateInputWidth",value:function(){if(this.mounted&&this.sizer&&void 0!==this.sizer.scrollWidth){var e=void 0;e=this.props.placeholder&&(!this.props.value||this.props.value&&this.props.placeholderIsMinWidth)?Math.max(this.sizer.scrollWidth,this.placeHolderSizer.scrollWidth)+2:this.sizer.scrollWidth+2,(e+="number"===this.props.type&&void 0===this.props.extraWidth?16:parseInt(this.props.extraWidth)||0)<this.props.minWidth&&(e=this.props.minWidth),e!==this.state.inputWidth&&this.setState({inputWidth:e})}}},{key:"getInput",value:function(){return this.input}},{key:"focus",value:function(){this.input.focus()}},{key:"blur",value:function(){this.input.blur()}},{key:"select",value:function(){this.input.select()}},{key:"renderStyles",value:function(){var e=this.props.injectStyles;return f&&e?u.default.createElement("style",{dangerouslySetInnerHTML:{__html:"input#"+this.state.inputId+"::-ms-clear {display: none;}"}}):null}},{key:"render",value:function(){var e=[this.props.defaultValue,this.props.value,""].reduce((function(e,t){return null!=e?e:t})),t=o({},this.props.style);t.display||(t.display="inline-block");var n=o({boxSizing:"content-box",width:this.state.inputWidth+"px"},this.props.inputStyle),r=function(e,t){var n={};for(var o in e)t.indexOf(o)>=0||Object.prototype.hasOwnProperty.call(e,o)&&(n[o]=e[o]);return n}(this.props,[]);return function(e){c.forEach((function(t){return delete e[t]}))}(r),r.className=this.props.inputClassName,r.id=this.state.inputId,r.style=n,u.default.createElement("div",{className:this.props.className,style:t},this.renderStyles(),u.default.createElement("input",o({},r,{ref:this.inputRef})),u.default.createElement("div",{ref:this.sizerRef,style:s},e),this.props.placeholder?u.default.createElement("div",{ref:this.placeHolderSizerRef,style:s},this.props.placeholder):null)}}]),t}(a.Component);_.propTypes={className:i.default.string,defaultValue:i.default.any,extraWidth:i.default.oneOfType([i.default.number,i.default.string]),id:i.default.string,injectStyles:i.default.bool,inputClassName:i.default.string,inputRef:i.default.func,inputStyle:i.default.object,minWidth:i.default.oneOfType([i.default.number,i.default.string]),onAutosize:i.default.func,onChange:i.default.func,placeholder:i.default.string,placeholderIsMinWidth:i.default.bool,style:i.default.object,value:i.default.any},_.defaultProps={minWidth:1,injectStyles:!0},t.default=_},216:(e,t,n)=>{"use strict";var o=n(138);function r(){}function a(){}a.resetWarningCache=r,e.exports=function(){function e(e,t,n,r,a,u){if(u!==o){var i=new Error("Calling PropTypes validators directly is not supported by the `prop-types` package. Use PropTypes.checkPropTypes() to call them. Read more at http://fb.me/use-check-prop-types");throw i.name="Invariant Violation",i}}function t(){return e}e.isRequired=e;var n={array:e,bool:e,func:e,number:e,object:e,string:e,symbol:e,any:e,arrayOf:t,element:e,elementType:e,instanceOf:t,node:e,objectOf:t,oneOf:t,oneOfType:t,shape:t,exact:t,checkPropTypes:a,resetWarningCache:r};return n.PropTypes=n,n}},863:(e,t,n)=>{e.exports=n(216)()},138:e=>{"use strict";e.exports="SECRET_DO_NOT_PASS_THIS_OR_YOU_WILL_BE_FIRED"},787:t=>{"use strict";t.exports=e}},n={};function o(e){var r=n[e];if(void 0!==r)return r.exports;var a=n[e]={exports:{}};return t[e].call(a.exports,a,a.exports,o),a.exports}o.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return o.d(t,{a:t}),t},o.d=(e,t)=>{for(var n in t)o.o(t,n)&&!o.o(e,n)&&Object.defineProperty(e,n,{enumerable:!0,get:t[n]})},o.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t),o.r=e=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})};var r={};return(()=>{"use strict";o.r(r),o.d(r,{JS__TOKEN__DELETE_BUTTON__CLASS_NAME:()=>n.JS__TOKEN__DELETE_BUTTON__CLASS_NAME,KEY_DOWN_HANDLER_CONFIG_OPTION:()=>n.KEY_DOWN_HANDLER_CONFIG_OPTION,default:()=>a});var e=o(922),t=o.n(e),n=o(754);const a=t()})(),r})()}));
 //# sourceMappingURL=index.js.map
